requirements.txt
faiss-cpu
groq
langchain-groq
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf
google-cloud-aiplatform>=1.38
langchain-core>=0.1.0